Initial Document Parsers
# Initial Document Parsers

Setting up the start of the DocumentParser implementations with a `DirectDocumentParser` which takes a `RawDocument` and parses it into a `Document`, with a direct mapping of `RawDocumentChunk`s to the `DocumentFragment`s.

This is especially useful for the LangChain file loaders which load a file into LangChain documents and convert directly to `RawDocument`s. For example, the CsvFileLoader loads a csv file from disk into one LangChain document per column; we convert that to a `RawDocument` with per-column chunks and the `DirectDocumentParser` will result in a `Document` with per-column fragments.

When we start loading files form other sources (e.g. loading a PDF from S3), we'll probably want to implement more powerful parsers for them -- e.g. we can add a PDFDocumentParser which will parse into fragments based on sections, etc.

```
ryanholinshead@Ryans-MacBook-Pro semantic-retrieval % npx jest directDocumentParser.test.ts
 PASS  __tests__/ingestion/document-parsers/directDocumentParser.test.ts
  DirectDocument parser
    ✓ parse single RawDocumentChunk into matching DocumentFragment (2 ms)
    ✓ parse multiple RawDocumentChunks into matching DocumentFragments

Test Suites: 1 passed, 1 total
Tests:       2 passed, 2 total
Snapshots:   0 total
Time:        0.519 s, estimated 1 s
```