Default FileSystem Data Source Implementation
# Default FileSystem Data Source Implementation

A basic default implementation for loading data from files (and directories of files) on disk. The implementation sets up a default mapping of file extensions to file loaders. Our default file loaders are just thin wrappers around the equivalent LangChain file loaders and can be configured using a param to the constructor