[LAS-429][RFC][WIP] callbacks
proof of concept. The idea is that every object in the flow (e.g. FileSystem) has
its own particular points at which callbacks may be called, and exposes that
particular list of points or "events" to the user. The user passes in callbacks as a
mapping from event name to callback.

A callback is an async (callback_input: object) => void. There is no obvious way to
do anything useful with return values, so it's void for now. Can rethink this.

callback_input is unstructured for the same reason every object has its own event
list: inputs are heterogenous. We can avoid imposing structure unless needed.

next steps:
- make sure we have a way to tie log points together with IDs. a request should have
a unique ID that gets passed to every callback
- move things up the object hierarchy and/or out to utils as appropriate.