[LAS-429][1/n] callbacks for TS flow
The idea is that every object in the flow (e.g. FileSystem) has
its own particular points at which callbacks may be called, and exposes that
particular list of points or "events" to the user. The user passes in callbacks as a
mapping from event name to callback.

A callback is an async (callback_input: object) => void. Return values, online agg,
and shared state generally not yet supported.

next steps:
- finish the list of classes / their public methods
- move things up the object hierarchy and/or out to utils as appropriate.