Make accessPassport optional for Retrieval
# Make accessPassport optional for Retrieval

Another one of the TODOs I was leaving for cleanup, but it came up in another PR so doing it now to prevent additional churn in future diffs.

Initial idea was to require AccessPassport so that we would have a more clear policy on accessing documents (no passport === no access). But that can be too restrictive for anyone using this library for general retrieval. So, we can still do the access checks and just let the policies fail/pass as they desire when no passport is provided.