Adding 10-ks and Running Ingestion Script
# Adding 10-ks and Running Ingestion Script

Added the 10-k markdown files and ran the ingestion script to get them into my pinecone DB. Note, cost was $0.28 for embeddings so it's not negligible when we re-run ingestion on all these docs.

For now, just reusing the .txt file loader but ideally we might:
- have a markdown loader that properly loads into fragments by heading-delineated sections
- or, convert these .md to txt and remove any clutter

Leaving as-is for now while I get the rest of the implementation done, but can revisit before demo if there's time.

See resulting metadataDB json here. Pinecone namespace for the run is `4de8f47d-4377-4f7b-9849-06dbd2e17e95` cc @jonathanlastmileai

Note this run also took over 10 minutes to create all embeddings and upsert to pinecone