Move ts files under /typescript
# Move ts files under /typescript

Move the typescript files (including tests, typescript examples, and config files) under the `/typescript` directory.

Noting that I had to do a trick with a `tsconfig.build.json`, following https://stackoverflow.com/a/61107618, in order for the test files to find jest types, while excluding test files from the compiled files in `/dist`.

Updated the github steps for build/test to account for the new directory