[RFC] FP monadic do for Result
I'd like to try this syntactic sugar for Result type in internal code.
It is very unpythonic and less readable until you get used to it,
but is commonplace in FP (Rust, Haskell...).

The basic idea is to abstract out statically-checked error handling.
This makes it very easy to compose steps each of which can error,
without manually checking if each one is an error.

Fully statically typed; pyright actually guarantees errors are handled.